fast_api_backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
fastapi==0.116.1
flatbuffers==25.2.10
fonttools==4.59.0
google-auth==2.40.3
google-genai==1.26.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
postgrest==1.1.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMatting==1.1.14
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-decouple==3.8
realtime==2.6.0
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
spotipy==2.25.1
starlette==0.47.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
tenacity==8.5.0
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
URLDecoder==1.0.3
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
wordcloud==1.9.4
